// projects/cloud-grid/src/enums/grid-topic.enum.ts
export enum GridTopicEnum {
    _RefreshQuery = '_RefreshQuery',
    _ColumnDefinition = '_ColumnDefinition',
    _ViewDefinition = '_ViewDefinition',
    _HistoryChange = '_HistoryChange',
    _History = '_History',
    _ListData = '_ListData',
    LinkFieldClick = 'LinkFieldClick',
    EnableRowOperation = 'EnableRowOperation',
    EnableFilterView = 'EnableFilterView',
    EnableColumnFrozen = 'EnableColumnFrozen',
    RowSelectMode = 'RowSelectMode',
    RowSelected = 'RowSelected',
    RowOperating = 'RowOperating',
    ColumnDefinition = 'ColumnDefinition',
    ViewDefinition = 'ViewDefinition',
    ListData = 'ListData',
    History = 'History',
    OpenAdvancePanel = 'OpenAdvancePanel',
    CloseAdvancePanel = 'CloseAdvancePanel',
    FilterViewCreateOrUpdate = 'FilterViewCreateOrUpdate',
    FreezeColumn = 'FreezeColumn',
    UnFreezeColumn = 'UnFreezeColumn',
    ColumnVisialSettingChange = 'ColumnVisialSettingChange'
}